Duracloud: Handle fetching many files from a space
refs #7481

Duracloud returns a maximum of 1000 files when browsing a space. Update
code to call that endpoint repeatedly to fetch all files with a given
prefix. To handle the larger number of files, convert _get_files_list to an
iterator.  Update browse to deal with paths one at a time, instead of
fetching all of them at once.  Update tests for new calls.